restructure to gain ability to supertest
